3 Term questions are implemented
Question object now takes a single parameter - set. This set is to be a
jagged array, [[operands],[operators],answer] which is generated in the
generateSum,Minus,etc functions.

The logic for creating a 3 term question is found in generateQuestion
which uses a 16 case switch to brute force the logic. If you got a
better way, please I need it :(